Move user controller into another folder, lock down routes
The user controller being in a "controllers" folder was risky, as it was being picked up by the Backend router if the correct URL was used. While it would throw an exception because it's not a true backend controller, it could've become a problem later if we changed how the Backend router worked since this controller gives unfettered access to users for testing purposes.

This controller now will never be picked up by the Backend router.

In addition, the routes provided by the Dusk service provider specifically require that the app environment is "dusk", so there's no chance they could be inadvertently available even in dev mode.